requirements.txt
# requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
django-ckeditor>=6.5.0
Pillow>=9.5.0
python-decouple>=3.8
beautifulsoup4
psycopg2-binary>=2.9.6  # For PostgreSQL
# or use mysql-connector-python for MySQL

# Optional but recommended
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0
celery>=5.2.0  # For async tasks
redis>=4.5.0  # For caching and Celery broker
gunicorn>=20.1.0  # For production deployment
whitenoise>=6.4.0  # For serving static files in production

# Development dependencies
pytest-django>=4.5.0
factory-boy>=3.2.0
coverage>=7.2.0